pyproject.toml
[project]
name = "money-printer"
version = "0.1.0"
description = "A-share multi-factor quantitative trading system"
requires-python = ">=3.10"
dependencies = [
    "akshare>=1.14",
    "pandas>=2.0",
    "numpy>=1.24",
    "polars>=0.20",
    "sqlalchemy>=2.0",
    "plotly>=5.18",
    "streamlit>=1.30",
    "loguru>=0.7",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "schedule>=1.2",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.4",
    "ipython>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mp"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
